MedVerifyApp/medverify-backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
tensorflow-cpu==2.16.2
opencv-python-headless==4.9.0.80
easyocr==1.7.1
google-cloud-bigquery==3.17.2
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.4
groq

pytest
pytest-flask
python-json-logger
